test(react-i18n): move detail type out of the file router folder
