+++ packages/java/tests/spring/react-i18n/src/main/frontend/types/detail.ts
