+++ packages/java/tests/spring/react-i18n/src/main/frontend/views/@index.tsx
@@ -1,6 +1,6 @@
 import type { ViewConfig } from '@vaadin/hilla-file-router/types.js';
 import { key, translate } from '@vaadin/hilla-react-i18n';
-import type { Detail } from './detail';
+import type { Detail } from '../types/detail.js';
 
 export const config: ViewConfig<Detail> = {
   title: key`viewtitle.home`,
